Add MyBooks, BookSearch, and Book components
Added MyBooks and BookSearch page components that will eventually be used to keep track of books
in state and to search for new books respectively. Also added Book component to display information
on an individual book. Added isOnBookshelf function to check whether a book is already in state and
added ternary operator to Book jsx to display either add or remove book button depending on result
of isOnBookshelf.